update: Web用コントローラにおける各メソッドへの例外処理の追加

+++ src/app/Http/Controllers/Web/PostController.php
@@ -80,31 +80,44 @@ class PostController extends Controller
      */
     public function store(StorePostRequest $request)
     {
-        $validated = $request->validated();
-
-        $post = new Post();
-        $post->title = $validated['title'];
-        $post->description = $validated['description'];
-        $post->level = $validated['level'];
-        $post->user_id = \Auth::id();
-        $post->text_id = $validated['text_id'];
-
-        if (isset($validated['file_name']))
+        try
         {
-            $file = $validated['file_name'];
-            $ext  = $file->getClientOriginalextension();
+            $validated = $request->validated();
 
-            // ファイルをストレージへ保存
-            $fileName = time() . '.' . $ext;
-            $file->storeAs('public/files', $fileName);
+            $post              = new Post();
+            $post->title       = $validated['title'];
+            $post->description = $validated['description'];
+            $post->level       = $validated['level'];
+            $post->user_id     = \Auth::id();
+            $post->text_id     = $validated['text_id'];
 
-            $post->file_name = $fileName;
-            $post->file_mimetype = $file->getMimeType();
-            $post->file_size = $file->getSize();
+            if (isset($validated['file_name']))
+            {
+                $file = $validated['file_name'];
+                $ext  = $file->getClientOriginalextension();
+
+                // ファイルをストレージへ保存
+                $fileName = time() . '.' . $ext;
+                $file->storeAs('public/files', $fileName);
+
+                $post->file_name     = $fileName;
+                $post->file_mimetype = $file->getMimeType();
+                $post->file_size     = $file->getSize();
+            }
+
+            if (!$post->save())
+            {
+                throw new \Exception('教案の投稿に失敗しました。時間を空けて再度投稿してください。');
+            }
+        }catch (\Exception $e)
+        {
+            return back()
+                ->withInput()
+                ->with('failureMessage', $e->getMessage());
         }
-        $post->save();
 
-        return redirect(route('myposts.index'))->with('successMessage', '教案を投稿しました。');
+        return redirect(route('myposts.index'))
+            ->with('successMessage', '教案を投稿しました。');
     }
 
     /**
@@ -163,33 +176,46 @@ class PostController extends Controller
     public function update(StorePostRequest $request, Post $post)
     {
         $this->authorize($post);
-        $validated = $request->validated();
 
-        // ファイルの差し替えを行う
-        if (isset($validated['file_name']))
+        try
         {
-            // 元ファイルの削除
-            \Storage::delete('public/files/' . $post->file_name);
+            $validated = $request->validated();
 
-            // 新ファイルの保存
-            $file = $validated['file_name'];
-            $ext  = $file->getClientOriginalextension();
-            $fileName = time() . '.' . $ext;
-            $file->storeAs('public/files', $fileName);
+            // ファイルの差し替えを行う
+            if (isset($validated['file_name']))
+            {
+                // 元ファイルの削除
+                \Storage::delete('public/files/' . $post->file_name);
 
-            // 新ファイルのデータをpostsテーブルの各カラムに保存
-            $post->file_name = $fileName;
-            $post->file_mimetype = $file->getMimeType();
-            $post->file_size = $file->getSize();
+                // 新ファイルの保存
+                $file     = $validated['file_name'];
+                $ext      = $file->getClientOriginalextension();
+                $fileName = time() . '.' . $ext;
+                $file->storeAs('public/files', $fileName);
+
+                // 新ファイルのデータをpostsテーブルの各カラムに保存
+                $post->file_name     = $fileName;
+                $post->file_mimetype = $file->getMimeType();
+                $post->file_size     = $file->getSize();
+            }
+
+            $post->title       = $validated['title'];
+            $post->description = $validated['description'];
+            $post->level       = $validated['level'];
+            $post->text_id     = $validated['text_id'];
+
+            if (!$post->save())
+            {
+                throw new \Exception('教案の編集に失敗しました。時間を空けて再度挑戦してください。');
+            }
+        } catch (\Exception $e)
+        {
+            return back()
+                ->withInput()
+                ->with('failureMessage', $e->getMessage());
         }
-
-        $post->title = $validated['title'];
-        $post->description = $validated['description'];
-        $post->level = $validated['level'];
-        $post->text_id = $validated['text_id'];
-        $post->save();
-
-        return redirect(route('posts.show', $post))->with('successMessage', '教案を更新しました。');
+        return redirect(route('posts.show', $post))
+                ->with('successMessage', '教案を更新しました。');
     }
 
     /**
@@ -201,12 +227,24 @@ class PostController extends Controller
     public function destroy(Post $post)
     {
         $this->authorize($post);
-        $post->delete();
-        // アップロードされたファイルの削除
-        \Storage::delete('public/files/' . $post->file_name);
 
-        return redirect(route('myposts.index'))->with('successMessage', '教案を削除しました。');
+        try
+        {
+            if (!$post->delete())
+            {
+                throw new \Exception('教案の削除に失敗しました。時間を空けて再度挑戦してください。');
+            }
 
+            // アップロードされたファイルの削除
+            \Storage::delete('public/files/' . $post->file_name);
+        } catch (\Exception $e)
+        {
+            return redirect(route('myposts.index'))
+                ->with('failureMessage', $e->getMessage());
+        }
+
+        return redirect(route('myposts.index'))
+            ->with('successMessage', '教案を削除しました。');
     }
 
     /**
